requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
xlrd>=2.0.1
scikit-learn>=1.3.0
statsmodels>=0.14.0
PyYAML>=6.0
tqdm>=4.65.0
joblib>=1.3.0
cvxpy>=1.3.0
cvxopt>=1.3.0
arch>=6.0.0
jupyter>=1.0.0
plotly>=5.14.0
